pyproject.toml
[tool.poetry]
name = "saucybot"
version = "0.1.0"
description = ""
authors = ["Sn0wCrack <snowcrack123@gmail.com>", "Goopypanther"]

[tool.poetry.dependencies]
python = "^3.6"
python-dotenv = "^0.10.3"
PixivPy = "^3.5"
"discord.py" = "^1.2"
requests-toolbelt = "^0.9.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
